+++ store/src/app/products/components/button/button.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { CartService } from '../cart/cart.service';
+import { CartService } from '../../services/cart.service';
 
 @Component({
   selector: 'app-button',
