+++ store/src/app/products/components/cart/cart.component.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/main-colors.scss";
+@import "../../../../scss/main-colors.scss";
 
 .container {
   width: 100%;
