+++ store/src/app/products/components/footer/footer.component.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/main-colors.scss";
+@import "../../../../scss/main-colors.scss";
 
 .section-products__link {
   display: block;
