+++ store/src/app/products/components/footer/footer.component.ts
