+++ store/src/app/products/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/main-colors.scss";
+@import "../../../../scss/main-colors.scss";
 
 .header {
   height: 101px;
