+++ store/src/app/products/components/header/header.component.ts
