+++ store/src/app/products/components/product-card/product-card.component.scss
