+++ store/src/app/products/components/product-info/product-info.component.scss
@@ -1,5 +1,3 @@
-@import "../../../scss/main-colors.scss";
-
 .product-info {
   padding: 0px 98px 260px 97px;
 }
