+++ store/src/app/products/components/product-info/product-info.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/interfaces/products.interface';
 import { ActivatedRoute } from '@angular/router';
-import { ProductsService } from '../products.service';
-import { CartService } from '../cart/cart.service';
+import { ProductsService } from '../../services/products.service';
+import { CartService } from '../../services/cart.service';
 
 @Component({
   selector: 'app-product-info',
