+++ store/src/app/products/components/tooltip/tooltip.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/interfaces/products.interface';
-import { CartService } from '../cart/cart.service';
+import { CartService } from '../../services/cart.service';
 
 @Component({
   selector: 'app-tooltip',
