+++ store/src/app/products/products-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { ProductInfoComponent } from './product-info/product-info.component';
+import { ProductInfoComponent } from './components/product-info/product-info.component';
 import { ProductsComponent } from './products.component';
-import { ProductsListComponent } from './products-list/products-list.component';
-import { CartComponent } from './cart/cart.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { ProductsListComponent } from './components/products-list/products-list.component';
+import { CartComponent } from './components/cart/cart.component';
+import { PageNotFoundComponent } from '../shared/components/page-not-found/page-not-found.component';
 
 const productsRoutes: Routes = [
   {
