+++ store/src/app/products/services/cart.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, OnInit } from '@angular/core';
 import { Product } from 'src/app/interfaces/products.interface';
-import { LocalStorageService } from '../local-storage.service';
+import { LocalStorageService } from './local-storage.service';
 
 @Injectable({
   providedIn: 'root'
