+++ store/src/app/products/services/products.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Product } from '../interfaces/products.interface';
+import { Product } from '../../interfaces/products.interface';
 
 @Injectable()
 export class ProductsService {
