+++ store/src/scss/main-colors.scss
@@ -5,3 +5,4 @@ $main-logo-bg: #d9d9d9;
 $product-button-bg: #2cfaa1;
 $product-button-bg-hover: #23d98a;
 $product-button-bg-active: #28fc24;
+$tooltip-bg: #8fbcef;
